Move xAI OAuth coverage tests to core utils
